release-notes-script/requirements.txt
atlassian-python-api==3.41.4
beautifulsoup4==4.12.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
Deprecated==1.2.14
idna==3.6
jmespath==1.0.1
numpy==1.26.3
oauthlib==3.2.2
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
tzdata==2023.4
urllib3==2.1.0
wrapt==1.16.0
